backend/requirements.txt
black==22.3.0
coverage==6.3.2
Django==4.0.3
gunicorn==20.1.0
mixer==7.2.2
mysqlclient==2.1.0
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dotenv==0.20.0
whitenoise==6.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0